server: add NVNC_DEGRADE=quality as an alternative to dropping frames
Congestion control has only ever had one actuator: hold quality fixed and
drop frames. For a desktop that is mostly still that is the right answer,
and it stays the default -- NVNC_DEGRADE is unset in every existing
deployment and nothing about their behaviour changes.

For a desktop showing video, or one whose main activity is scrolling a
dense plot, it is the wrong answer twice over. The obvious way is that a
slideshow at pristine quality is worse than smooth motion at a quality
that fits. The less obvious way is a feedback loop: each dropped frame
widens the displacement between the frames that survive, which pushes it
past the encoder's motion search, which makes those frames dramatically
more expensive, which causes more dropping.

Setting NVNC_DEGRADE=quality drives the encoder's bitrate ceiling from
the bandwidth estimate instead. Frame dropping stays in place underneath
as the latency bound and takes over once the ceiling reaches its floor.

The estimate cannot be used as the ceiling directly. It measures what was
delivered, which on an idle desktop is almost nothing, so treating it as
the available bandwidth would throttle a fast link for being asked to
carry little. The in-flight backlog distinguishes the two: the ceiling
only descends when there is a queue to show for it, and otherwise climbs
back towards what the resolution warrants. Backing off is allowed once a
second and probing forward once every five, because every change rebuilds
the encoder and costs a key frame.

Measured against a 1280x720 scrolling noise plot at 30 Hz, with the
client reading at a fixed rate so that TCP applies real backpressure:

    link        framerate (default)   quality
    100 Mb/s    30.0 fps              30.0 fps
    5 Mb/s       2.0 fps              30.0 fps
    2 Mb/s       6.9 fps              30.0 fps
    1 Mb/s      17.2 fps              29.2 fps
    0.5 Mb/s     8.4 fps              20.9 fps

At 0.5 Mb/s the ceiling bottoms out at its floor and frame dropping takes
over, which is the intended handoff. At 100 Mb/s the two modes are
indistinguishable, which is the point: this costs nothing when it is not
needed.

congestion-server.c and rfb-throttle-client.py are added to reproduce the
table. The client speaks the Fence and ContinuousUpdates pseudo-encodings,
without which the server's estimator is never fed at all.

+++ src/enc/h264/open-h264.c
@@ -261,7 +261,7 @@ static void open_h264_set_max_bitrate(struct encoder* enc, int max_bitrate)
 	 * control mode and always warrants a rebuild.
 	 */
 	if (self->max_bitrate != 0 && max_bitrate != 0) {
-		double change = fabs(max_bitrate - self->max_bitrate) /
+		double change = fabs((double)max_bitrate - self->max_bitrate) /
 			(double)self->max_bitrate;
 		if (change < BITRATE_HYSTERESIS)
 			return;
